feat: add eval function and modify main
Introduced an evaluation function, 'eval', in new file eval.go which makes
the program capable of doing mathematical operations and handle symbolic
expressions. The main function was modified to use the 'eval' function.

The values `PI` and `E` and a function `pow(a, b)` have been predefined in
the evaluator as multiple statements aren't supported yet.