feat: implement block expressions and multiple statements in parser
Code was modified throughout the project to introduce block expressions
in the AST (Abstract Syntax Tree) layer and support multiple statements
in the parser. The feature was added to facilitate the execution of
multiple sequential calculations, defined by semicolon-separated
statements, within a single block, enhancing the usability and
expressive power of the calculator.

E.g., the following was made possible:

```
PI = 3.14159265358979323846
E = 2.71828182845904523536
pow(PI, 2) + pow(E, 2);
```

The lexer was adjusted to recognize the semicolon as a token type. The
syntax for defining block expressions was introduced in the parser.
Existing code for execution (eval) and printing were adjusted to
handle block expressions correctly.

The method modifying global constants `PI` and `E` was removed from the
eval initialization function and moved to the main execution example in
order to demonstrate the new option of defining constants within the
input block.